merge from origin source

+++ scripts/get_chromedriver.sh
@@ -5,12 +5,13 @@ if [ ! -d $chromedriver_dir ]; then
   mkdir $chromedriver_dir
 fi
 
-if [[ $os_name == 'Linux' ]]; then
-  cd chromedriver  && curl -L https://chromedriver.storage.googleapis.com/2.29/chromedriver_linux64.zip > tmp.zip &&  unzip -o tmp.zip && rm tmp.zip
+echo "downloading chromedriver"
+
+if [[ $os_name == 'Linux' && ! -f $chromedriver_dir/chromedriver ]]; then
+  cd chromedriver  && curl -L https://chromedriver.storage.googleapis.com/2.29/chromedriver_linux64.zip > tmp.zip &&  unzip -o tmp.zip && rm tmp.zip 
   # wait until download finish
   sleep 5
-elif [[ $os_name == 'Darwin' ]]; then
-  cd chromedriver  &&  curl -L https://chromedriver.storage.googleapis.com/2.29/chromedriver_mac64.zip  | tar xz
-  # wait until download finish
+elif [[ $os_name == 'Darwin' && ! -f $chromedriver_dir/chromedriver ]]; then
+  cd chromedriver  &&  curl -L https://chromedriver.storage.googleapis.com/2.29/chromedriver_mac64.zip  | tar xz 
   sleep 5
 fi
